@@ -8,48 +8,49 @@
     "build": "next build",
     "serve": "next start",
     "lint": "next lint",
-    "format": "prettier --write . '!(node_module|.next|contracts)/**/*'",
+    "format": "prettier --write . '!(node_modules|.next|contracts)/**/*'",
     "check-types": "tsc --noEmit --incremental",
     "vercel": "vercel",
     "vercel:yolo": "vercel --build-env NEXT_PUBLIC_IGNORE_BUILD_ERROR=true"
   },
   "dependencies": {
-    "@ethersproject/networks": "^5.7.1",
-    "@ethersproject/web": "^5.7.1",
+    "@ethersproject/providers": "^5.7.2",
     "@heroicons/react": "^2.0.11",
-    "@rainbow-me/rainbowkit": "^0.11.0",
-    "@uniswap/sdk": "^3.0.3",
-    "daisyui": "^2.31.0",
-    "ethers": "^5.0.0",
-    "next": "^13.1.6",
-    "nextjs-progressbar": "^0.0.16",
+    "@rainbow-me/rainbowkit": "1.3.0",
+    "@uniswap/sdk-core": "^4.0.1",
+    "@uniswap/v2-sdk": "^3.0.1",
+    "blo": "^1.0.1",
+    "daisyui": "^4.4.19",
+    "next": "^14.0.4",
+    "nprogress": "^0.2.0",
+    "qrcode.react": "^3.1.0",
     "react": "^18.2.0",
-    "react-blockies": "^1.4.1",
     "react-copy-to-clipboard": "^5.1.0",
     "react-dom": "^18.2.0",
-    "react-fast-marquee": "^1.3.5",
     "react-hot-toast": "^2.4.0",
     "use-debounce": "^8.0.4",
     "usehooks-ts": "^2.7.2",
-    "wagmi": "^0.11.6",
+    "viem": "1.19.9",
+    "wagmi": "1.4.12",
     "zustand": "^4.1.2"
   },
   "devDependencies": {
     "@trivago/prettier-plugin-sort-imports": "^4.1.1",
     "@types/node": "^17.0.35",
+    "@types/nprogress": "^0",
     "@types/react": "^18.0.9",
-    "@types/react-blockies": "^1.4.1",
     "@types/react-copy-to-clipboard": "^5.0.4",
     "@typescript-eslint/eslint-plugin": "^5.39.0",
     "autoprefixer": "^10.4.12",
     "eslint": "^8.15.0",
-    "eslint-config-next": "^13.1.6",
+    "eslint-config-next": "^14.0.4",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
     "postcss": "^8.4.16",
     "prettier": "^2.8.4",
-    "tailwindcss": "^3.1.8",
-    "typescript": "^4.9.5",
-    "vercel": "^28.15.1"
+    "tailwindcss": "^3.3.3",
+    "type-fest": "^4.6.0",
+    "typescript": "^5.1.6",
+    "vercel": "^32.4.1"
   }
 }
