@@ -1,17 +1,29 @@
-import { useAccountBalance } from "~~/hooks/scaffold-eth";
-import { getTargetNetwork } from "~~/utils/scaffold-eth";
+"use client";
 
-type TBalanceProps = {
-  address?: string;
+import { useState } from "react";
+import { Address } from "viem";
+import { useAccountBalance } from "~~/hooks/scaffold-eth";
+import { useTargetNetwork } from "~~/hooks/scaffold-eth/useTargetNetwork";
+
+type BalanceProps = {
+  address?: Address;
   className?: string;
+  usdMode?: boolean;
 };
 
 /**
  * Display (ETH & USD) balance of an ETH address.
  */
-export const Balance = ({ address, className = "" }: TBalanceProps) => {
-  const configuredNetwork = getTargetNetwork();
-  const { balance, price, isError, isLoading, onToggleBalance, isEthBalance } = useAccountBalance(address);
+export const Balance = ({ address, className = "", usdMode }: BalanceProps) => {
+  const { targetNetwork } = useTargetNetwork();
+  const { balance, price, isError, isLoading } = useAccountBalance(address);
+  const [displayUsdMode, setDisplayUsdMode] = useState(price > 0 ? Boolean(usdMode) : false);
+
+  const toggleBalanceMode = () => {
+    if (price > 0) {
+      setDisplayUsdMode(prevMode => !prevMode);
+    }
+  };
 
   if (!address || isLoading || balance === null) {
     return (
@@ -35,18 +47,18 @@ export const Balance = ({ address, className = "" }: TBalanceProps) => {
   return (
     <button
       className={`btn btn-sm btn-ghost flex flex-col font-normal items-center hover:bg-transparent ${className}`}
-      onClick={onToggleBalance}
+      onClick={toggleBalanceMode}
     >
       <div className="w-full flex items-center justify-center">
-        {isEthBalance ? (
-          <>
-            <span>{balance?.toFixed(4)}</span>
-            <span className="text-[0.8em] font-bold ml-1">{configuredNetwork.nativeCurrency.symbol}</span>
-          </>
-        ) : (
+        {displayUsdMode ? (
           <>
             <span className="text-[0.8em] font-bold mr-1">$</span>
             <span>{(balance * price).toFixed(2)}</span>
+          </>
+        ) : (
+          <>
+            <span>{balance?.toFixed(4)}</span>
+            <span className="text-[0.8em] font-bold ml-1">{targetNetwork.nativeCurrency.symbol}</span>
           </>
         )}
       </div>
