@@ -1,13 +1,20 @@
+"use client";
+
 import { useState } from "react";
-import { ethers } from "ethers";
+import { createWalletClient, http, parseEther } from "viem";
+import { hardhat } from "viem/chains";
 import { useAccount, useNetwork } from "wagmi";
-import { hardhat, localhost } from "wagmi/chains";
 import { BanknotesIcon } from "@heroicons/react/24/outline";
 import { useAccountBalance, useTransactor } from "~~/hooks/scaffold-eth";
-import { getLocalProvider } from "~~/utils/scaffold-eth";
 
 // Number of ETH faucet sends to an address
 const NUM_OF_ETH = "1";
+const FAUCET_ADDRESS = "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266";
+
+const localWalletClient = createWalletClient({
+  chain: hardhat,
+  transport: http(),
+});
 
 /**
  * FaucetButton button which lets you grab eth.
@@ -15,16 +22,22 @@ const NUM_OF_ETH = "1";
 export const FaucetButton = () => {
   const { address } = useAccount();
   const { balance } = useAccountBalance(address);
+
   const { chain: ConnectedChain } = useNetwork();
+
   const [loading, setLoading] = useState(false);
-  const provider = getLocalProvider(localhost);
-  const signer = provider?.getSigner();
-  const faucetTxn = useTransactor(signer);
+
+  const faucetTxn = useTransactor(localWalletClient);
 
   const sendETH = async () => {
     try {
       setLoading(true);
-      await faucetTxn({ to: address, value: ethers.utils.parseEther(NUM_OF_ETH) });
+      await faucetTxn({
+        chain: hardhat,
+        account: FAUCET_ADDRESS,
+        to: address,
+        value: parseEther(NUM_OF_ETH),
+      });
       setLoading(false);
     } catch (error) {
       console.error("⚡️ ~ file: FaucetButton.tsx:sendETH ~ error", error);
@@ -33,7 +46,7 @@ export const FaucetButton = () => {
   };
 
   // Render only on local chain
-  if (!ConnectedChain || ConnectedChain.id !== hardhat.id) {
+  if (ConnectedChain?.id !== hardhat.id) {
     return null;
   }
 
@@ -41,19 +54,17 @@ export const FaucetButton = () => {
     <div
       className={
         balance
-          ? ""
-          : "tooltip tooltip-bottom tooltip-secondary tooltip-open font-bold before:left-auto before:transform-none before:content-[attr(data-tip)] before:right-0"
+          ? "ml-1"
+          : "ml-1 tooltip tooltip-bottom tooltip-secondary tooltip-open font-bold before:left-auto before:transform-none before:content-[attr(data-tip)] before:right-0"
       }
       data-tip="Grab funds from faucet"
     >
-      <button
-        className={`btn btn-secondary btn-sm px-2 rounded-full ${
-          loading ? "loading before:!w-4 before:!h-4 before:!mx-0" : ""
-        }`}
-        onClick={sendETH}
-        disabled={loading}
-      >
-        {!loading && <BanknotesIcon className="h-4 w-4" />}
+      <button className="btn btn-secondary btn-sm px-2 rounded-full" onClick={sendETH} disabled={loading}>
+        {!loading ? (
+          <BanknotesIcon className="h-4 w-4" />
+        ) : (
+          <span className="loading loading-spinner loading-xs"></span>
+        )}
       </button>
     </div>
   );
