@@ -1,34 +1,41 @@
 import { useCallback, useEffect, useState } from "react";
-import { isAddress } from "ethers/lib/utils";
-import Blockies from "react-blockies";
+import { blo } from "blo";
+import { useDebounce } from "usehooks-ts";
+import { Address, isAddress } from "viem";
 import { useEnsAddress, useEnsAvatar, useEnsName } from "wagmi";
-import { CommonInputProps, InputBase } from "~~/components/scaffold-eth";
-
-// ToDo:  move this function to an utility file
-const isENS = (address = "") => address.endsWith(".eth") || address.endsWith(".xyz");
+import { CommonInputProps, InputBase, isENS } from "~~/components/scaffold-eth";
 
 /**
  * Address input with ENS name resolution
  */
-export const AddressInput = ({ value, name, placeholder, onChange }: CommonInputProps) => {
+export const AddressInput = ({ value, name, placeholder, onChange, disabled }: CommonInputProps<Address | string>) => {
+  // Debounce the input to keep clean RPC calls when resolving ENS names
+  // If the input is an address, we don't need to debounce it
+  const _debouncedValue = useDebounce(value, 500);
+  const debouncedValue = isAddress(value) ? value : _debouncedValue;
+  const isDebouncedValueLive = debouncedValue === value;
+
+  // If the user changes the input after an ENS name is already resolved, we want to remove the stale result
+  const settledValue = isDebouncedValueLive ? debouncedValue : undefined;
+
   const { data: ensAddress, isLoading: isEnsAddressLoading } = useEnsAddress({
-    name: value,
-    enabled: isENS(value),
+    name: settledValue,
+    enabled: isENS(debouncedValue),
     chainId: 1,
     cacheTime: 30_000,
   });
 
   const [enteredEnsName, setEnteredEnsName] = useState<string>();
   const { data: ensName, isLoading: isEnsNameLoading } = useEnsName({
-    address: value,
-    enabled: isAddress(value),
+    address: settledValue as Address,
+    enabled: isAddress(debouncedValue),
     chainId: 1,
     cacheTime: 30_000,
   });
 
   const { data: ensAvatar } = useEnsAvatar({
-    address: value,
-    enabled: isAddress(value),
+    name: ensName,
+    enabled: Boolean(ensName),
     chainId: 1,
     cacheTime: 30_000,
   });
@@ -38,12 +45,12 @@ export const AddressInput = ({ value, name, placeholder, onChange }: CommonInput
     if (!ensAddress) return;
 
     // ENS resolved successfully
-    setEnteredEnsName(value);
+    setEnteredEnsName(debouncedValue);
     onChange(ensAddress);
-  }, [ensAddress, onChange, value]);
+  }, [ensAddress, onChange, debouncedValue]);
 
   const handleChange = useCallback(
-    (newValue: string) => {
+    (newValue: Address) => {
       setEnteredEnsName(undefined);
       onChange(newValue);
     },
@@ -51,13 +58,13 @@ export const AddressInput = ({ value, name, placeholder, onChange }: CommonInput
   );
 
   return (
-    <InputBase
+    <InputBase<Address>
       name={name}
       placeholder={placeholder}
       error={ensAddress === null}
-      value={value}
+      value={value as Address}
       onChange={handleChange}
-      disabled={isEnsAddressLoading || isEnsNameLoading}
+      disabled={isEnsAddressLoading || isEnsNameLoading || disabled}
       prefix={
         ensName && (
           <div className="flex bg-base-300 rounded-l-full items-center">
@@ -73,7 +80,11 @@ export const AddressInput = ({ value, name, placeholder, onChange }: CommonInput
           </div>
         )
       }
-      suffix={value && <Blockies className="!rounded-full" seed={value?.toLowerCase() as string} size={7} scale={5} />}
+      suffix={
+        // Don't want to use nextJS Image here (and adding remote patterns for the URL)
+        // eslint-disable-next-line @next/next/no-img-element
+        value && <img alt="" className="!rounded-full" src={blo(value as `0x${string}`)} width="35" height="35" />
+      }
     />
   );
 };
