@@ -1,17 +1,13 @@
 import { useCallback } from "react";
-import { ethers } from "ethers";
+import { hexToString, isHex, stringToHex } from "viem";
 import { CommonInputProps, InputBase } from "~~/components/scaffold-eth";
 
-export const Bytes32Input = ({ value, onChange, name, placeholder }: CommonInputProps) => {
+export const Bytes32Input = ({ value, onChange, name, placeholder, disabled }: CommonInputProps) => {
   const convertStringToBytes32 = useCallback(() => {
     if (!value) {
       return;
     }
-    onChange(
-      ethers.utils.isHexString(value)
-        ? ethers.utils.parseBytes32String(value)
-        : ethers.utils.formatBytes32String(value),
-    );
+    onChange(isHex(value) ? hexToString(value, { size: 32 }) : stringToHex(value, { size: 32 }));
   }, [onChange, value]);
 
   return (
@@ -20,6 +16,7 @@ export const Bytes32Input = ({ value, onChange, name, placeholder }: CommonInput
       value={value}
       placeholder={placeholder}
       onChange={onChange}
+      disabled={disabled}
       suffix={
         <div
           className="self-center cursor-pointer text-xl font-semibold px-4 text-accent"
