@@ -1,14 +1,10 @@
 import { useCallback } from "react";
-import { ethers } from "ethers";
+import { bytesToString, isHex, toBytes, toHex } from "viem";
 import { CommonInputProps, InputBase } from "~~/components/scaffold-eth";
 
-export const BytesInput = ({ value, onChange, name, placeholder }: CommonInputProps) => {
+export const BytesInput = ({ value, onChange, name, placeholder, disabled }: CommonInputProps) => {
   const convertStringToBytes = useCallback(() => {
-    onChange(
-      ethers.utils.isHexString(value)
-        ? ethers.utils.toUtf8String(value)
-        : ethers.utils.hexlify(ethers.utils.toUtf8Bytes(value)),
-    );
+    onChange(isHex(value) ? bytesToString(toBytes(value)) : toHex(toBytes(value)));
   }, [onChange, value]);
 
   return (
@@ -17,6 +13,7 @@ export const BytesInput = ({ value, onChange, name, placeholder }: CommonInputPr
       value={value}
       placeholder={placeholder}
       onChange={onChange}
+      disabled={disabled}
       suffix={
         <div
           className="self-center cursor-pointer text-xl font-semibold px-4 text-accent"
