@@ -3,12 +3,11 @@ import { CommonInputProps } from "~~/components/scaffold-eth";
 
 type InputBaseProps<T> = CommonInputProps<T> & {
   error?: boolean;
-  disabled?: boolean;
   prefix?: ReactNode;
   suffix?: ReactNode;
 };
 
-export const InputBase = <T extends { toString: () => string } = string>({
+export const InputBase = <T extends { toString: () => string } | undefined = string>({
   name,
   value,
   onChange,
@@ -36,7 +35,7 @@ export const InputBase = <T extends { toString: () => string } = string>({
     <div className={`flex border-2 border-base-300 bg-base-200 rounded-full text-accent ${modifier}`}>
       {prefix}
       <input
-        className="input input-ghost focus:outline-none focus:bg-transparent focus:text-gray-400 h-[2.2rem] min-h-[2.2rem] px-4 border w-full font-medium placeholder:text-accent/50 text-gray-400"
+        className="input input-ghost focus-within:border-transparent focus:outline-none focus:bg-transparent focus:text-gray-400 h-[2.2rem] min-h-[2.2rem] px-4 border w-full font-medium placeholder:text-accent/50 text-gray-400"
         placeholder={placeholder}
         name={name}
         value={value?.toString()}
