@@ -1,9 +1,9 @@
 import { useCallback, useEffect, useState } from "react";
-import { BigNumber, ethers } from "ethers";
 import { CommonInputProps, InputBase, IntegerVariant, isValidInteger } from "~~/components/scaffold-eth";
 
-type IntegerInputProps = CommonInputProps<string | BigNumber> & {
+type IntegerInputProps = CommonInputProps<string | bigint> & {
   variant?: IntegerVariant;
+  disableMultiplyBy1e18?: boolean;
 };
 
 export const IntegerInput = ({
@@ -11,14 +11,19 @@ export const IntegerInput = ({
   onChange,
   name,
   placeholder,
+  disabled,
   variant = IntegerVariant.UINT256,
+  disableMultiplyBy1e18 = false,
 }: IntegerInputProps) => {
   const [inputError, setInputError] = useState(false);
   const multiplyBy1e18 = useCallback(() => {
     if (!value) {
       return;
     }
-    onChange(ethers.utils.parseEther(value.toString()));
+    if (typeof value === "bigint") {
+      return onChange(value * 10n ** 18n);
+    }
+    return onChange(BigInt(Math.round(Number(value) * 10 ** 18)));
   }, [onChange, value]);
 
   useEffect(() => {
@@ -36,13 +41,19 @@ export const IntegerInput = ({
       placeholder={placeholder}
       error={inputError}
       onChange={onChange}
+      disabled={disabled}
       suffix={
-        !inputError && (
+        !inputError &&
+        !disableMultiplyBy1e18 && (
           <div
             className="space-x-4 flex tooltip tooltip-top tooltip-secondary before:content-[attr(data-tip)] before:right-[-10px] before:left-auto before:transform-none"
             data-tip="Multiply by 10^18 (wei)"
           >
-            <button className="cursor-pointer font-semibold px-4 text-accent" onClick={multiplyBy1e18}>
+            <button
+              className={`${disabled ? "cursor-not-allowed" : "cursor-pointer"} font-semibold px-4 text-accent`}
+              onClick={multiplyBy1e18}
+              disabled={disabled}
+            >
               ∗
             </button>
           </div>
