@@ -1,6 +1,5 @@
 export * from "./useAccountBalance";
 export * from "./useAnimationConfig";
-export * from "./useAutoConnect";
 export * from "./useBurnerWallet";
 export * from "./useDeployedContractInfo";
 export * from "./useNativeCurrencyPrice";
@@ -14,3 +13,4 @@ export * from "./useScaffoldEventHistory";
 export * from "./useTransactor";
 export * from "./useFetchBlocks";
 export * from "./useContractLogs";
+export * from "./useAutoConnect";
