@@ -1,12 +1,14 @@
 import { useCallback, useEffect, useState } from "react";
+import { useTargetNetwork } from "./useTargetNetwork";
+import { Address } from "viem";
 import { useBalance } from "wagmi";
 import { useGlobalState } from "~~/services/store/store";
-import { getTargetNetwork } from "~~/utils/scaffold-eth";
 
-export function useAccountBalance(address?: string) {
+export function useAccountBalance(address?: Address) {
   const [isEthBalance, setIsEthBalance] = useState(true);
   const [balance, setBalance] = useState<number | null>(null);
   const price = useGlobalState(state => state.nativeCurrencyPrice);
+  const { targetNetwork } = useTargetNetwork();
 
   const {
     data: fetchedBalanceData,
@@ -15,7 +17,7 @@ export function useAccountBalance(address?: string) {
   } = useBalance({
     address,
     watch: true,
-    chainId: getTargetNetwork().id,
+    chainId: targetNetwork.id,
   });
 
   const onToggleBalance = useCallback(() => {
@@ -28,7 +30,7 @@ export function useAccountBalance(address?: string) {
     if (fetchedBalanceData?.formatted) {
       setBalance(Number(fetchedBalanceData.formatted));
     }
-  }, [fetchedBalanceData]);
+  }, [fetchedBalanceData, targetNetwork]);
 
   return { balance, price, isError, isLoading, onToggleBalance, isEthBalance };
 }
