@@ -1,33 +1,42 @@
-import { useEffect } from "react";
-import { useEffectOnce, useLocalStorage } from "usehooks-ts";
+import { useEffectOnce, useLocalStorage, useReadLocalStorage } from "usehooks-ts";
+import { Chain, hardhat } from "viem/chains";
 import { Connector, useAccount, useConnect } from "wagmi";
-import { hardhat } from "wagmi/chains";
 import scaffoldConfig from "~~/scaffold.config";
-import { burnerWalletId, defaultBurnerChainId } from "~~/services/web3/wagmi-burner/BurnerConnector";
-import { getTargetNetwork } from "~~/utils/scaffold-eth";
+import { burnerWalletId } from "~~/services/web3/wagmi-burner/BurnerConnector";
+import { getTargetNetworks } from "~~/utils/scaffold-eth";
 
-const walletIdStorageKey = "scaffoldEth2.wallet";
+const SCAFFOLD_WALLET_STORAGE_KEY = "scaffoldEth2.wallet";
+const WAGMI_WALLET_STORAGE_KEY = "wagmi.wallet";
+
+// ID of the SAFE connector instance
+const SAFE_ID = "safe";
 
 /**
  * This function will get the initial wallet connector (if any), the app will connect to
+ * @param initialNetwork
  * @param previousWalletId
  * @param connectors
  * @returns
  */
 const getInitialConnector = (
+  initialNetwork: Chain,
   previousWalletId: string,
-  connectors: Connector<any, any, any>[],
+  connectors: Connector[],
 ): { connector: Connector | undefined; chainId?: number } | undefined => {
-  const burnerConfig = scaffoldConfig.burnerWallet;
-  const targetNetwork = getTargetNetwork();
+  // Look for the SAFE connector instance and connect to it instantly if loaded in SAFE frame
+  const safeConnectorInstance = connectors.find(connector => connector.id === SAFE_ID && connector.ready);
 
-  const allowBurner = burnerConfig.enabled && (burnerConfig.onlyLocal ? targetNetwork.id === hardhat.id : true);
+  if (safeConnectorInstance) {
+    return { connector: safeConnectorInstance };
+  }
+
+  const allowBurner = scaffoldConfig.onlyLocalBurnerWallet ? initialNetwork.id === hardhat.id : true;
 
   if (!previousWalletId) {
     // The user was not connected to a wallet
     if (allowBurner && scaffoldConfig.walletAutoConnect) {
       const connector = connectors.find(f => f.id === burnerWalletId);
-      return { connector, chainId: defaultBurnerChainId };
+      return { connector, chainId: initialNetwork.id };
     }
   } else {
     // the user was connected to wallet
@@ -48,23 +57,21 @@ const getInitialConnector = (
  * Automatically connect to a wallet/connector based on config and prior wallet
  */
 export const useAutoConnect = (): void => {
-  const [walletId, setWalletId] = useLocalStorage<string>(walletIdStorageKey, "");
+  const wagmiWalletValue = useReadLocalStorage<string>(WAGMI_WALLET_STORAGE_KEY);
+  const [walletId, setWalletId] = useLocalStorage<string>(SCAFFOLD_WALLET_STORAGE_KEY, wagmiWalletValue ?? "");
   const connectState = useConnect();
-  const accountState = useAccount();
-
-  useEffect(() => {
-    if (accountState.isConnected) {
-      // user is connected, set walletName
-      setWalletId(accountState.connector?.id ?? "");
-    } else {
-      // user has disconnected, reset walletName
+  useAccount({
+    onConnect({ connector }) {
+      setWalletId(connector?.id ?? "");
+    },
+    onDisconnect() {
+      window.localStorage.setItem(WAGMI_WALLET_STORAGE_KEY, JSON.stringify(""));
       setWalletId("");
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [accountState.isConnected, accountState.connector?.name]);
+    },
+  });
 
   useEffectOnce(() => {
-    const initialConnector = getInitialConnector(walletId, connectState.connectors);
+    const initialConnector = getInitialConnector(getTargetNetworks()[0], walletId, connectState.connectors);
 
     if (initialConnector?.connector) {
       connectState.connect({ connector: initialConnector.connector, chainId: initialConnector.chainId });
