@@ -1,158 +1,139 @@
-import { useCallback, useEffect, useRef } from "react";
-import { BytesLike, Signer, Wallet, ethers } from "ethers";
-import { useDebounce } from "use-debounce";
+import { useCallback, useEffect, useRef, useState } from "react";
 import { useLocalStorage } from "usehooks-ts";
-import { useProvider } from "wagmi";
+import { Chain, Hex, HttpTransport, PrivateKeyAccount, createWalletClient, http } from "viem";
+import { generatePrivateKey, privateKeyToAccount } from "viem/accounts";
+import { WalletClient, usePublicClient } from "wagmi";
 
 const burnerStorageKey = "scaffoldEth2.burnerWallet.sk";
 
 /**
- * Is the private key valid
- * @internal
- * @param pk
- * @returns
+ * Checks if the private key is valid
  */
-const isValidSk = (pk: BytesLike | undefined | null): boolean => {
+const isValidSk = (pk: Hex | string | undefined | null): boolean => {
   return pk?.length === 64 || pk?.length === 66;
 };
 
 /**
- * If no burner is found in localstorage, we will use a new default wallet
+ * If no burner is found in localstorage, we will generate a random private key
  */
-const newDefaultWallet = ethers.Wallet.createRandom();
+const newDefaultPrivateKey = generatePrivateKey();
 
 /**
- * Save the current burner private key from storage
- * Can be used outside of react.  Used by the burnerConnector.
- * @internal
- * @returns
+ * Save the current burner private key to local storage
  */
-export const saveBurnerSK = (wallet: Wallet): void => {
+export const saveBurnerSK = (privateKey: Hex): void => {
   if (typeof window != "undefined" && window != null) {
-    window?.localStorage?.setItem(burnerStorageKey, wallet.privateKey);
+    window?.localStorage?.setItem(burnerStorageKey, privateKey);
   }
 };
 
 /**
- * Gets the current burner private key from storage
- * Can be used outside of react.  Used by the burnerConnector.
- * @internal
- * @returns
+ * Gets the current burner private key from local storage
  */
-export const loadBurnerSK = (): string => {
-  let currentSk = "";
+export const loadBurnerSK = (): Hex => {
+  let currentSk: Hex = "0x";
   if (typeof window != "undefined" && window != null) {
-    currentSk = window?.localStorage?.getItem?.(burnerStorageKey)?.replaceAll('"', "") ?? "";
+    currentSk = (window?.localStorage?.getItem?.(burnerStorageKey)?.replaceAll('"', "") ?? "0x") as Hex;
   }
 
   if (!!currentSk && isValidSk(currentSk)) {
     return currentSk;
   } else {
-    saveBurnerSK(newDefaultWallet);
-    return newDefaultWallet.privateKey;
+    saveBurnerSK(newDefaultPrivateKey);
+    return newDefaultPrivateKey;
   }
 };
 
-/**
- * #### Summary
- * Return type of useBurnerSigner:
- *
- * ##### ✏️ Notes
- * - provides signer
- * - methods of interacting with burner signer
- * - methods to save and loadd signer from local storage
- *
- * @category Hooks
- */
-export type TBurnerSigner = {
-  signer: Signer | undefined;
-  account: string | undefined;
-  /**
-   * create a new burner signer
-   */
+type BurnerAccount = {
+  walletClient: WalletClient | undefined;
+  account: PrivateKeyAccount | undefined;
+  // creates a new burner account
   generateNewBurner: () => void;
-  /**
-   * explictly save burner to storage
-   */
+  // explicitly save burner to storage
   saveBurner: () => void;
 };
 
 /**
- * #### Summary
- * A hook that creates a burner signer/address and provides ways of interacting with
- * and updating the signer
- *
- * @category Hooks
- *
- * @param localProvider localhost provider
- * @returns IBurnerSigner
+ * Creates a burner wallet
  */
-export const useBurnerWallet = (): TBurnerSigner => {
-  const [burnerSk, setBurnerSk] = useLocalStorage<BytesLike>(burnerStorageKey, newDefaultWallet.privateKey);
+export const useBurnerWallet = (): BurnerAccount => {
+  const [burnerSk, setBurnerSk] = useLocalStorage<Hex>(burnerStorageKey, newDefaultPrivateKey);
 
-  const provider = useProvider();
-  const walletRef = useRef<Wallet>();
+  const publicClient = usePublicClient();
+  const [walletClient, setWalletClient] = useState<WalletClient<HttpTransport, Chain, PrivateKeyAccount>>();
+  const [generatedPrivateKey, setGeneratedPrivateKey] = useState<Hex>("0x");
+  const [account, setAccount] = useState<PrivateKeyAccount>();
   const isCreatingNewBurnerRef = useRef(false);
 
-  const [signer] = useDebounce(walletRef.current, 200, {
-    trailing: true,
-    equalityFn: (a, b) => a?.address === b?.address && a != null && b != null,
-  });
-  const account = walletRef.current?.address;
-
-  /**
-   * callback to save current wallet sk
-   */
   const saveBurner = useCallback(() => {
-    setBurnerSk(walletRef.current?.privateKey ?? "");
-  }, [setBurnerSk]);
+    setBurnerSk(generatedPrivateKey);
+  }, [setBurnerSk, generatedPrivateKey]);
 
-  /**
-   * create a new burnerkey
-   */
   const generateNewBurner = useCallback(() => {
-    if (provider && !isCreatingNewBurnerRef.current) {
+    if (publicClient && !isCreatingNewBurnerRef.current) {
       console.log("🔑 Create new burner wallet...");
       isCreatingNewBurnerRef.current = true;
 
-      const wallet = Wallet.createRandom().connect(provider);
-      setBurnerSk(() => {
-        console.log("🔥 ...Save new burner wallet");
-        isCreatingNewBurnerRef.current = false;
-        return wallet.privateKey;
+      const randomPrivateKey = generatePrivateKey();
+      const randomAccount = privateKeyToAccount(randomPrivateKey);
+
+      const client = createWalletClient({
+        chain: publicClient.chain,
+        account: randomAccount,
+        transport: http(),
       });
-      return wallet;
+
+      setWalletClient(client);
+      setGeneratedPrivateKey(randomPrivateKey);
+      setAccount(randomAccount);
+
+      setBurnerSk(() => {
+        console.log("🔥 Saving new burner wallet");
+        isCreatingNewBurnerRef.current = false;
+        return randomPrivateKey;
+      });
+      return client;
     } else {
       console.log("⚠ Could not create burner wallet");
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [provider?.network?.chainId]);
+  }, [publicClient.chain.id]);
 
   /**
    * Load wallet with burnerSk
    * connect and set wallet, once we have burnerSk and valid provider
    */
   useEffect(() => {
-    if (burnerSk && provider.network.chainId) {
-      let wallet: Wallet | undefined = undefined;
+    if (burnerSk && publicClient.chain.id) {
+      let wallet: WalletClient<HttpTransport, Chain, PrivateKeyAccount> | undefined = undefined;
       if (isValidSk(burnerSk)) {
-        wallet = new ethers.Wallet(burnerSk, provider);
+        const randomAccount = privateKeyToAccount(burnerSk);
+
+        wallet = createWalletClient({
+          chain: publicClient.chain,
+          account: randomAccount,
+          transport: http(),
+        });
+
+        setGeneratedPrivateKey(burnerSk);
+        setAccount(randomAccount);
       } else {
-        wallet = generateNewBurner?.();
+        wallet = generateNewBurner();
       }
 
       if (wallet == null) {
         throw "Error:  Could not create burner wallet";
       }
-      walletRef.current = wallet;
-      saveBurner?.();
+
+      setWalletClient(wallet);
+      saveBurner();
     }
 
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [burnerSk, provider?.network?.chainId]);
+  }, [burnerSk, publicClient.chain.id]);
 
   return {
-    signer,
+    walletClient,
     account,
     generateNewBurner,
     saveBurner,
