@@ -1,38 +1,37 @@
 import { useEffect, useState } from "react";
-import { ethers } from "ethers";
-import { useProvider } from "wagmi";
+import { Address, Log } from "viem";
+import { usePublicClient } from "wagmi";
 
-export const useContractLogs = (address: string) => {
-  const [logs, setLogs] = useState<ethers.providers.Log[]>([]);
-  const provider = useProvider();
+export const useContractLogs = (address: Address) => {
+  const [logs, setLogs] = useState<Log[]>([]);
+  const client = usePublicClient();
 
   useEffect(() => {
     const fetchLogs = async () => {
       try {
-        const filter = {
+        const existingLogs = await client.getLogs({
           address: address,
-          fromBlock: 0,
+          fromBlock: 0n,
           toBlock: "latest",
-        };
-        const existingLogs = await provider.getLogs(filter);
+        });
         setLogs(existingLogs);
       } catch (error) {
         console.error("Failed to fetch logs:", error);
       }
     };
-
-    const handleLog = (log: ethers.providers.Log) => {
-      setLogs(prevLogs => [...prevLogs, log]);
-    };
-
-    const filter = { address: address };
-
     fetchLogs();
-    provider.on(filter, handleLog);
-    return () => {
-      provider.off(filter, handleLog);
-    };
-  }, [address, provider]);
+
+    return client.watchBlockNumber({
+      onBlockNumber: async (blockNumber, prevBlockNumber) => {
+        const newLogs = await client.getLogs({
+          address: address,
+          fromBlock: prevBlockNumber,
+          toBlock: "latest",
+        });
+        setLogs(prevLogs => [...prevLogs, ...newLogs]);
+      },
+    });
+  }, [address, client]);
 
   return logs;
 };
