@@ -1,20 +1,19 @@
 import { useEffect, useState } from "react";
+import { useTargetNetwork } from "./useTargetNetwork";
 import { useIsMounted } from "usehooks-ts";
-import { useProvider } from "wagmi";
-import scaffoldConfig from "~~/scaffold.config";
+import { usePublicClient } from "wagmi";
 import { Contract, ContractCodeStatus, ContractName, contracts } from "~~/utils/scaffold-eth/contract";
 
 /**
- * Gets the matching contract info from the contracts file generated by `yarn deploy`
- * @param contractName - name of deployed contract
+ * Gets the matching contract info for the provided contract name from the contracts present in deployedContracts.ts
+ * and externalContracts.ts corresponding to targetNetworks configured in scaffold.config.ts
  */
 export const useDeployedContractInfo = <TContractName extends ContractName>(contractName: TContractName) => {
   const isMounted = useIsMounted();
-  const deployedContract = contracts?.[scaffoldConfig.targetNetwork.id]?.[0]?.contracts?.[
-    contractName as ContractName
-  ] as Contract<TContractName>;
+  const { targetNetwork } = useTargetNetwork();
+  const deployedContract = contracts?.[targetNetwork.id]?.[contractName as ContractName] as Contract<TContractName>;
   const [status, setStatus] = useState<ContractCodeStatus>(ContractCodeStatus.LOADING);
-  const provider = useProvider({ chainId: scaffoldConfig.targetNetwork.id });
+  const publicClient = usePublicClient({ chainId: targetNetwork.id });
 
   useEffect(() => {
     const checkContractDeployment = async () => {
@@ -22,7 +21,9 @@ export const useDeployedContractInfo = <TContractName extends ContractName>(cont
         setStatus(ContractCodeStatus.NOT_FOUND);
         return;
       }
-      const code = await provider.getCode((deployedContract as Contract<TContractName>).address);
+      const code = await publicClient.getBytecode({
+        address: deployedContract.address,
+      });
 
       if (!isMounted()) {
         return;
@@ -36,7 +37,7 @@ export const useDeployedContractInfo = <TContractName extends ContractName>(cont
     };
 
     checkContractDeployment();
-  }, [isMounted, contractName, deployedContract, provider]);
+  }, [isMounted, contractName, deployedContract, publicClient]);
 
   return {
     data: status === ContractCodeStatus.DEPLOYED ? deployedContract : undefined,
