@@ -1,41 +1,52 @@
 import { useCallback, useEffect, useState } from "react";
-import { ethers } from "ethers";
-import { localhost } from "wagmi/chains";
+import {
+  Block,
+  Hash,
+  Transaction,
+  TransactionReceipt,
+  createTestClient,
+  publicActions,
+  walletActions,
+  webSocket,
+} from "viem";
+import { hardhat } from "viem/chains";
 import { decodeTransactionData } from "~~/utils/scaffold-eth";
-import { getLocalProvider } from "~~/utils/scaffold-eth";
-import { Block } from "~~/utils/scaffold-eth/block";
 
 const BLOCKS_PER_PAGE = 20;
 
-const provider = getLocalProvider(localhost) || new ethers.providers.JsonRpcProvider("http://localhost:8545");
+export const testClient = createTestClient({
+  chain: hardhat,
+  mode: "hardhat",
+  transport: webSocket("ws://127.0.0.1:8545"),
+})
+  .extend(publicActions)
+  .extend(walletActions);
 
 export const useFetchBlocks = () => {
   const [blocks, setBlocks] = useState<Block[]>([]);
   const [transactionReceipts, setTransactionReceipts] = useState<{
-    [key: string]: ethers.providers.TransactionReceipt;
+    [key: string]: TransactionReceipt;
   }>({});
   const [currentPage, setCurrentPage] = useState(0);
-  const [totalBlocks, setTotalBlocks] = useState(0);
-  const [isLoading, setIsLoading] = useState(false);
+  const [totalBlocks, setTotalBlocks] = useState(0n);
   const [error, setError] = useState<Error | null>(null);
 
   const fetchBlocks = useCallback(async () => {
-    setIsLoading(true);
     setError(null);
 
     try {
-      const blockNumber = await provider.getBlockNumber();
+      const blockNumber = await testClient.getBlockNumber();
       setTotalBlocks(blockNumber);
 
-      const startingBlock = blockNumber - currentPage * BLOCKS_PER_PAGE;
+      const startingBlock = blockNumber - BigInt(currentPage * BLOCKS_PER_PAGE);
       const blockNumbersToFetch = Array.from(
-        { length: Math.min(BLOCKS_PER_PAGE, startingBlock + 1) },
-        (_, i) => startingBlock - i,
+        { length: Number(BLOCKS_PER_PAGE < startingBlock + 1n ? BLOCKS_PER_PAGE : startingBlock + 1n) },
+        (_, i) => startingBlock - BigInt(i),
       );
 
       const blocksWithTransactions = blockNumbersToFetch.map(async blockNumber => {
         try {
-          return provider.getBlockWithTransactions(blockNumber);
+          return testClient.getBlock({ blockNumber, includeTransactions: true });
         } catch (err) {
           setError(err instanceof Error ? err : new Error("An error occurred."));
           throw err;
@@ -44,15 +55,15 @@ export const useFetchBlocks = () => {
       const fetchedBlocks = await Promise.all(blocksWithTransactions);
 
       fetchedBlocks.forEach(block => {
-        block.transactions.forEach(tx => decodeTransactionData(tx));
+        block.transactions.forEach(tx => decodeTransactionData(tx as Transaction));
       });
 
       const txReceipts = await Promise.all(
         fetchedBlocks.flatMap(block =>
           block.transactions.map(async tx => {
             try {
-              const receipt = await provider.getTransactionReceipt(tx.hash);
-              return { [tx.hash]: receipt };
+              const receipt = await testClient.getTransactionReceipt({ hash: (tx as Transaction).hash });
+              return { [(tx as Transaction).hash]: receipt };
             } catch (err) {
               setError(err instanceof Error ? err : new Error("An error occurred."));
               throw err;
@@ -66,7 +77,6 @@ export const useFetchBlocks = () => {
     } catch (err) {
       setError(err instanceof Error ? err : new Error("An error occurred."));
     }
-    setIsLoading(false);
   }, [currentPage]);
 
   useEffect(() => {
@@ -74,42 +84,43 @@ export const useFetchBlocks = () => {
   }, [fetchBlocks]);
 
   useEffect(() => {
-    const handleNewBlock = async (blockNumber: number) => {
+    const handleNewBlock = async (newBlock: any) => {
       try {
-        const newBlock = await provider.getBlockWithTransactions(blockNumber);
-        if (!blocks.some(block => block.number === newBlock.number)) {
-          if (currentPage === 0) {
-            setBlocks(prevBlocks => [newBlock, ...prevBlocks.slice(0, BLOCKS_PER_PAGE - 1)]);
-
-            newBlock.transactions.forEach(tx => decodeTransactionData(tx));
-
-            const receipts = await Promise.all(
-              newBlock.transactions.map(async tx => {
-                try {
-                  const receipt = await provider.getTransactionReceipt(tx.hash);
-                  return { [tx.hash]: receipt };
-                } catch (err) {
-                  setError(err instanceof Error ? err : new Error("An error occurred."));
-                  throw err;
-                }
-              }),
+        if (currentPage === 0) {
+          if (newBlock.transactions.length > 0) {
+            const transactionsDetails = await Promise.all(
+              newBlock.transactions.map((txHash: string) => testClient.getTransaction({ hash: txHash as Hash })),
             );
-
-            setTransactionReceipts(prevReceipts => ({ ...prevReceipts, ...Object.assign({}, ...receipts) }));
+            newBlock.transactions = transactionsDetails;
           }
-          setTotalBlocks(blockNumber + 1);
+
+          newBlock.transactions.forEach((tx: Transaction) => decodeTransactionData(tx as Transaction));
+
+          const receipts = await Promise.all(
+            newBlock.transactions.map(async (tx: Transaction) => {
+              try {
+                const receipt = await testClient.getTransactionReceipt({ hash: (tx as Transaction).hash });
+                return { [(tx as Transaction).hash]: receipt };
+              } catch (err) {
+                setError(err instanceof Error ? err : new Error("An error occurred fetching receipt."));
+                throw err;
+              }
+            }),
+          );
+
+          setBlocks(prevBlocks => [newBlock, ...prevBlocks.slice(0, BLOCKS_PER_PAGE - 1)]);
+          setTransactionReceipts(prevReceipts => ({ ...prevReceipts, ...Object.assign({}, ...receipts) }));
+        }
+        if (newBlock.number) {
+          setTotalBlocks(newBlock.number);
         }
       } catch (err) {
         setError(err instanceof Error ? err : new Error("An error occurred."));
       }
     };
 
-    provider.on("block", handleNewBlock);
-
-    return () => {
-      provider.off("block", handleNewBlock);
-    };
-  }, [blocks, currentPage]);
+    return testClient.watchBlocks({ onBlock: handleNewBlock, includeTransactions: true });
+  }, [currentPage]);
 
   return {
     blocks,
@@ -117,7 +128,6 @@ export const useFetchBlocks = () => {
     currentPage,
     totalBlocks,
     setCurrentPage,
-    isLoading,
     error,
   };
 };
