@@ -1,19 +1,23 @@
 import { useState } from "react";
+import { useTargetNetwork } from "./useTargetNetwork";
 import { Abi, ExtractAbiFunctionNames } from "abitype";
-import { utils } from "ethers";
 import { useContractWrite, useNetwork } from "wagmi";
-import { getParsedEthersError } from "~~/components/scaffold-eth";
 import { useDeployedContractInfo, useTransactor } from "~~/hooks/scaffold-eth";
-import { getTargetNetwork, notification } from "~~/utils/scaffold-eth";
+import { getParsedError, notification } from "~~/utils/scaffold-eth";
 import { ContractAbi, ContractName, UseScaffoldWriteConfig } from "~~/utils/scaffold-eth/contract";
 
+type UpdatedArgs = Parameters<ReturnType<typeof useContractWrite<Abi, string, undefined>>["writeAsync"]>[0];
+
 /**
- * @dev wrapper for wagmi's useContractWrite hook(with config prepared by usePrepareContractWrite hook) which loads in deployed contract abi and address automatically
+ * Wrapper around wagmi's useContractWrite hook which automatically loads (by name) the contract ABI and address from
+ * the contracts present in deployedContracts.ts & externalContracts.ts corresponding to targetNetworks configured in scaffold.config.ts
  * @param config - The config settings, including extra wagmi configuration
- * @param config.contractName - deployed contract name
+ * @param config.contractName - contract name
  * @param config.functionName - name of the function to be called
  * @param config.args - arguments for the function
  * @param config.value - value in ETH that will be sent with transaction
+ * @param config.blockConfirmations - number of block confirmations to wait for (default: 1)
+ * @param config.onBlockConfirmation - callback that will be called after blockConfirmations.
  */
 export const useScaffoldContractWrite = <
   TContractName extends ContractName,
@@ -31,65 +35,55 @@ export const useScaffoldContractWrite = <
   const { chain } = useNetwork();
   const writeTx = useTransactor();
   const [isMining, setIsMining] = useState(false);
-  const configuredNetwork = getTargetNetwork();
-
-  const { overrides, ...restConfig } = writeConfig;
+  const { targetNetwork } = useTargetNetwork();
 
   const wagmiContractWrite = useContractWrite({
-    mode: "recklesslyUnprepared",
-    chainId: configuredNetwork.id,
+    chainId: targetNetwork.id,
     address: deployedContractData?.address,
     abi: deployedContractData?.abi as Abi,
-    args: args as unknown[],
     functionName: functionName as any,
-    overrides: {
-      value: value ? utils.parseEther(value) : undefined,
-      ...overrides,
-    },
-    ...restConfig,
+    args: args as unknown[],
+    value: value,
+    ...writeConfig,
   });
 
   const sendContractWriteTx = async ({
-    args,
-    value,
-    overrides,
+    args: newArgs,
+    value: newValue,
+    ...otherConfig
   }: {
     args?: UseScaffoldWriteConfig<TContractName, TFunctionName>["args"];
     value?: UseScaffoldWriteConfig<TContractName, TFunctionName>["value"];
-    overrides?: UseScaffoldWriteConfig<TContractName, TFunctionName>["overrides"];
-  } = {}) => {
+  } & UpdatedArgs = {}) => {
     if (!deployedContractData) {
-      notification.error("Target Contract is not deployed, did you forgot to run `yarn deploy`?");
+      notification.error("Target Contract is not deployed, did you forget to run `yarn deploy`?");
       return;
     }
     if (!chain?.id) {
       notification.error("Please connect your wallet");
       return;
     }
-    if (chain?.id !== configuredNetwork.id) {
-      notification.error("You on the wrong network");
+    if (chain?.id !== targetNetwork.id) {
+      notification.error("You are on the wrong network");
       return;
     }
 
     if (wagmiContractWrite.writeAsync) {
       try {
         setIsMining(true);
-        await writeTx(
-          wagmiContractWrite.writeAsync({
-            recklesslySetUnpreparedArgs: args as unknown[],
-            recklesslySetUnpreparedOverrides:
-              value && overrides
-                ? { value: utils.parseEther(value), ...overrides }
-                : value
-                ? { value: utils.parseEther(value) }
-                : overrides
-                ? overrides
-                : undefined,
-          }),
+        const writeTxResult = await writeTx(
+          () =>
+            wagmiContractWrite.writeAsync({
+              args: newArgs ?? args,
+              value: newValue ?? value,
+              ...otherConfig,
+            }),
           { onBlockConfirmation, blockConfirmations },
         );
+
+        return writeTxResult;
       } catch (e: any) {
-        const message = getParsedEthersError(e);
+        const message = getParsedError(e);
         notification.error(message);
       } finally {
         setIsMining(false);
