@@ -1,35 +1,38 @@
+import { useTargetNetwork } from "./useTargetNetwork";
 import { Abi, ExtractAbiEventNames } from "abitype";
+import { Log } from "viem";
 import { useContractEvent } from "wagmi";
-import { useDeployedContractInfo } from "~~/hooks/scaffold-eth";
-import { getTargetNetwork } from "~~/utils/scaffold-eth";
-import { AbiEventArgs, ContractAbi, ContractName, UseScaffoldEventConfig } from "~~/utils/scaffold-eth/contract";
+import { addIndexedArgsToEvent, useDeployedContractInfo } from "~~/hooks/scaffold-eth";
+import { ContractAbi, ContractName, UseScaffoldEventConfig } from "~~/utils/scaffold-eth/contract";
 
 /**
- * @dev wrapper for wagmi's useContractEvent
+ * Wrapper around wagmi's useEventSubscriber hook which automatically loads (by name) the contract ABI and
+ * address from the contracts present in deployedContracts.ts & externalContracts.ts
  * @param config - The config settings
  * @param config.contractName - deployed contract name
  * @param config.eventName - name of the event to listen for
- * @param config.listener - the callback that receives event
- * @param config.once - if set to true it will receive only a single event, then stop listening for the event. Defaults to false
+ * @param config.listener - the callback that receives events. If only interested in 1 event, call `unwatch` inside of the listener
  */
 export const useScaffoldEventSubscriber = <
   TContractName extends ContractName,
   TEventName extends ExtractAbiEventNames<ContractAbi<TContractName>>,
-  TEventInputs extends AbiEventArgs<ContractAbi<TContractName>, TEventName> & any[],
 >({
   contractName,
   eventName,
   listener,
-  once,
-}: UseScaffoldEventConfig<TContractName, TEventName, TEventInputs>) => {
+}: UseScaffoldEventConfig<TContractName, TEventName>) => {
   const { data: deployedContractData } = useDeployedContractInfo(contractName);
+  const { targetNetwork } = useTargetNetwork();
+
+  const addIndexedArgsToLogs = (logs: Log[]) => logs.map(addIndexedArgsToEvent);
+  const listenerWithIndexedArgs = (logs: Log[]) =>
+    listener(addIndexedArgsToLogs(logs) as Parameters<typeof listener>[0]);
 
   return useContractEvent({
     address: deployedContractData?.address,
     abi: deployedContractData?.abi as Abi,
-    chainId: getTargetNetwork().id,
-    listener: listener as (...args: unknown[]) => void,
-    eventName: eventName as string,
-    once: once ?? false,
+    chainId: targetNetwork.id,
+    listener: listenerWithIndexedArgs,
+    eventName,
   });
 };
