@@ -1,22 +1,20 @@
-import * as chains from "wagmi/chains";
+import * as chains from "viem/chains";
 
 export type ScaffoldConfig = {
-  targetNetwork: chains.Chain;
+  targetNetworks: readonly chains.Chain[];
   pollingInterval: number;
   alchemyApiKey: string;
-  burnerWallet: {
-    enabled: boolean;
-    onlyLocal: boolean;
-  };
+  walletConnectProjectId: string;
+  onlyLocalBurnerWallet: boolean;
   walletAutoConnect: boolean;
 };
 
 const scaffoldConfig = {
-  // The network where your DApp lives in
-  targetNetwork: chains.hardhat,
+  // The networks on which your DApp is live
+  targetNetworks: [chains.hardhat],
 
   // The interval at which your front-end polls the RPC servers for new data
-  // it has no effect on the local network
+  // it has no effect if you only target the local network (default is 4000)
   pollingInterval: 30000,
 
   // This is ours Alchemy's default API key.
@@ -25,13 +23,14 @@ const scaffoldConfig = {
   // .env.local for local testing, and in the Vercel/system env config for live apps.
   alchemyApiKey: process.env.NEXT_PUBLIC_ALCHEMY_API_KEY || "oKxs-03sij-U_N0iOlrSsZFr29-IqbuF",
 
-  // Burner Wallet configuration
-  burnerWallet: {
-    // Set it to false to completely remove burner wallet from all networks
-    enabled: true,
-    // Only show the Burner Wallet when running on hardhat network
-    onlyLocal: true,
-  },
+  // This is ours WalletConnect's default project ID.
+  // You can get your own at https://cloud.walletconnect.com
+  // It's recommended to store it in an env variable:
+  // .env.local for local testing, and in the Vercel/system env config for live apps.
+  walletConnectProjectId: process.env.NEXT_PUBLIC_WALLET_CONNECT_PROJECT_ID || "3a8170812b534d0ff9d794f19a901d64",
+
+  // Only show the Burner Wallet when running on hardhat network
+  onlyLocalBurnerWallet: true,
 
   /**
    * Auto connect:
@@ -39,6 +38,6 @@ const scaffoldConfig = {
    * 2. If user is not connected to any wallet:  On reload, connect to burner wallet if burnerWallet.enabled is true && burnerWallet.onlyLocal is false
    */
   walletAutoConnect: true,
-} satisfies ScaffoldConfig;
+} as const satisfies ScaffoldConfig;
 
 export default scaffoldConfig;
