@@ -1,4 +1,6 @@
-import create from "zustand";
+import { create } from "zustand";
+import scaffoldConfig from "~~/scaffold.config";
+import { ChainWithAttributes } from "~~/utils/scaffold-eth";
 
 /**
  * Zustand Store
@@ -9,12 +11,16 @@ import create from "zustand";
  * Think about it as a global useState.
  */
 
-type TGlobalState = {
+type GlobalState = {
   nativeCurrencyPrice: number;
   setNativeCurrencyPrice: (newNativeCurrencyPriceState: number) => void;
+  targetNetwork: ChainWithAttributes;
+  setTargetNetwork: (newTargetNetwork: ChainWithAttributes) => void;
 };
 
-export const useGlobalState = create<TGlobalState>(set => ({
+export const useGlobalState = create<GlobalState>(set => ({
   nativeCurrencyPrice: 0,
   setNativeCurrencyPrice: (newValue: number): void => set(() => ({ nativeCurrencyPrice: newValue })),
+  targetNetwork: scaffoldConfig.targetNetworks[0],
+  setTargetNetwork: (newTargetNetwork: ChainWithAttributes) => set(() => ({ targetNetwork: newTargetNetwork })),
 }));
