@@ -1,14 +1,13 @@
 import { StaticJsonRpcProvider } from "@ethersproject/providers";
-import { Wallet } from "ethers";
-import { Chain, Connector } from "wagmi";
+import { Address, Chain, HttpTransport, PrivateKeyAccount, WalletClient, createWalletClient, http } from "viem";
+import { privateKeyToAccount } from "viem/accounts";
+import { Connector } from "wagmi";
 import { loadBurnerSK } from "~~/hooks/scaffold-eth";
 import { BurnerConnectorError, BurnerConnectorErrorList } from "~~/services/web3/wagmi-burner/BurnerConnectorErrors";
 import { BurnerConnectorData, BurnerConnectorOptions } from "~~/services/web3/wagmi-burner/BurnerConnectorTypes";
-import { getTargetNetwork } from "~~/utils/scaffold-eth";
 
 export const burnerWalletId = "burner-wallet";
 export const burnerWalletName = "Burner Wallet";
-export const defaultBurnerChainId = getTargetNetwork().id;
 
 /**
  * This class is a wagmi connector for BurnerWallet.  Its used by {@link burnerWalletConfig}
@@ -19,10 +18,9 @@ export class BurnerConnector extends Connector<StaticJsonRpcProvider, BurnerConn
   readonly ready = true;
 
   private provider?: StaticJsonRpcProvider;
-  /**
-   * this is the store for getWallet()
-   */
-  private burnerWallet: Wallet | undefined;
+
+  // store for getWallet()
+  private burnerWallet: WalletClient<HttpTransport, Chain, PrivateKeyAccount> | undefined;
 
   constructor(config: { chains?: Chain[]; options: BurnerConnectorOptions }) {
     super(config);
@@ -35,6 +33,21 @@ export class BurnerConnector extends Connector<StaticJsonRpcProvider, BurnerConn
       this.provider = new StaticJsonRpcProvider(chain.rpcUrls.default.http[0]);
     }
     return this.provider;
+  }
+
+  async getWalletClient(config?: { chainId?: number | undefined } | undefined) {
+    const chain = this.getChainFromId(config?.chainId);
+    if (!this.burnerWallet) {
+      const burnerAccount = privateKeyToAccount(loadBurnerSK());
+
+      const client = createWalletClient({
+        chain: chain,
+        account: burnerAccount,
+        transport: http(),
+      });
+      this.burnerWallet = client;
+    }
+    return Promise.resolve(this.burnerWallet);
   }
 
   async connect(config?: { chainId?: number | undefined } | undefined): Promise<Required<BurnerConnectorData>> {
@@ -60,8 +73,9 @@ export class BurnerConnector extends Connector<StaticJsonRpcProvider, BurnerConn
       provider: this.provider,
     };
 
-    return data;
+    return Promise.resolve(data);
   }
+
   private getChainFromId(chainId?: number) {
     const resolveChainId = chainId ?? this.options.defaultChainId;
     const chain = this.chains.find(f => f.id === resolveChainId);
@@ -76,10 +90,9 @@ export class BurnerConnector extends Connector<StaticJsonRpcProvider, BurnerConn
     return Promise.resolve();
   }
 
-  async getAccount(): Promise<string> {
-    const wallet = this.getWallet();
-    const account = wallet.address;
-    return account;
+  async getAccount(): Promise<Address> {
+    const bunerAccount = privateKeyToAccount(loadBurnerSK());
+    return bunerAccount.address as Address;
   }
 
   async getChainId(): Promise<number> {
@@ -92,16 +105,6 @@ export class BurnerConnector extends Connector<StaticJsonRpcProvider, BurnerConn
     return Promise.resolve(chainId);
   }
 
-  async getSigner(): Promise<any> {
-    const account = await this.getAccount();
-    const signer = this.getWallet();
-
-    if (signer == null || (await signer.getAddress()) !== account) {
-      throw new BurnerConnectorError(BurnerConnectorErrorList.signerNotResolved);
-    }
-
-    return Promise.resolve(signer);
-  }
   async isAuthorized() {
     try {
       const account = await this.getAccount();
@@ -111,19 +114,41 @@ export class BurnerConnector extends Connector<StaticJsonRpcProvider, BurnerConn
     }
   }
 
-  private getWallet(): Wallet {
-    if (this.burnerWallet == null) {
-      this.burnerWallet = new Wallet(loadBurnerSK(), this.provider);
-    }
-    return this.burnerWallet;
+  protected async onAccountsChanged() {
+    const chainId = await this.getChainId();
+    const chain = this.getChainFromId(chainId);
+    const bunerAccount = privateKeyToAccount(loadBurnerSK());
+
+    const client = createWalletClient({
+      chain: chain,
+      account: bunerAccount,
+      transport: http(),
+    });
+    this.burnerWallet = client;
   }
 
-  protected onAccountsChanged(): void {
-    this.burnerWallet = new Wallet(loadBurnerSK(), this.provider);
+  async switchChain(chainId: number) {
+    const chain = this.getChainFromId(chainId);
+    this.provider = new StaticJsonRpcProvider(chain.rpcUrls.default.http[0]);
+
+    await this.onChainChanged();
+    return chain;
   }
-  protected onChainChanged(): void {
-    this.burnerWallet = new Wallet(loadBurnerSK(), this.provider);
+
+  protected async onChainChanged() {
+    const chainId = await this.getChainId();
+    const chain = this.getChainFromId(chainId);
+    const bunerAccount = privateKeyToAccount(loadBurnerSK());
+
+    const client = createWalletClient({
+      chain: chain,
+      account: bunerAccount,
+      transport: http(),
+    });
+    this.burnerWallet = client;
+    this.emit("change", { chain: { id: chainId, unsupported: false } });
   }
+
   protected onDisconnect(error: Error): void {
     if (error) console.warn(error);
   }
