@@ -5,15 +5,14 @@ export const BurnerConnectorErrorList = {
   accountNotFound: "Account not found",
   couldNotConnect: "Could not connect to network",
   unsupportedBurnerChain: "This network is not supported for burner connector",
-  chainIdNotResolved: "Cound not resolve chainId",
-  signerNotResolved: "Cound not resolve signer",
+  chainIdNotResolved: "Could not resolve chainId",
   chainNotSupported: "Chain is not supported, check burner wallet config",
 } as const;
 
 /**
  * A union of all the BurnerConnectorErrorList
  */
-export type BurnerConnectorErrorTypes = typeof BurnerConnectorErrorList[keyof typeof BurnerConnectorErrorList];
+export type BurnerConnectorErrorTypes = (typeof BurnerConnectorErrorList)[keyof typeof BurnerConnectorErrorList];
 
 export class BurnerConnectorError extends Error {
   constructor(errorType: BurnerConnectorErrorTypes, message?: string) {
