@@ -5,22 +5,24 @@ import {
   ledgerWallet,
   metaMaskWallet,
   rainbowWallet,
+  safeWallet,
   walletConnectWallet,
 } from "@rainbow-me/rainbowkit/wallets";
+import * as chains from "viem/chains";
 import { configureChains } from "wagmi";
-import * as chains from "wagmi/chains";
 import { alchemyProvider } from "wagmi/providers/alchemy";
 import { publicProvider } from "wagmi/providers/public";
 import scaffoldConfig from "~~/scaffold.config";
 import { burnerWalletConfig } from "~~/services/web3/wagmi-burner/burnerWalletConfig";
-import { getTargetNetwork } from "~~/utils/scaffold-eth";
+import { getTargetNetworks } from "~~/utils/scaffold-eth";
 
-const configuredNetwork = getTargetNetwork();
-const burnerConfig = scaffoldConfig.burnerWallet;
+const targetNetworks = getTargetNetworks();
+const { onlyLocalBurnerWallet } = scaffoldConfig;
 
 // We always want to have mainnet enabled (ENS resolution, ETH price, etc). But only once.
-const enabledChains =
-  (configuredNetwork.id as number) === 1 ? [configuredNetwork] : [configuredNetwork, chains.mainnet];
+const enabledChains = targetNetworks.find(network => network.id === 1)
+  ? targetNetworks
+  : [...targetNetworks, chains.mainnet];
 
 /**
  * Chains for the app
@@ -30,14 +32,14 @@ export const appChains = configureChains(
   [
     alchemyProvider({
       apiKey: scaffoldConfig.alchemyApiKey,
-      priority: 0,
     }),
-    publicProvider({ priority: 1 }),
+    publicProvider(),
   ],
   {
+    // We might not need this checkout https://github.com/scaffold-eth/scaffold-eth-2/pull/45#discussion_r1024496359, will test and remove this before merging
     stallTimeout: 3_000,
-    // Sets pollingInterval if using chain's other than local hardhat chain
-    ...(configuredNetwork.id !== chains.hardhat.id
+    // Sets pollingInterval if using chains other than local hardhat chain
+    ...(targetNetworks.find(network => network.id !== chains.hardhat.id)
       ? {
           pollingInterval: scaffoldConfig.pollingInterval,
         }
@@ -45,13 +47,22 @@ export const appChains = configureChains(
   },
 );
 
+const walletsOptions = { chains: appChains.chains, projectId: scaffoldConfig.walletConnectProjectId };
 const wallets = [
-  metaMaskWallet({ chains: appChains.chains, shimDisconnect: true }),
-  walletConnectWallet({ chains: appChains.chains }),
-  ledgerWallet({ chains: appChains.chains }),
-  braveWallet({ chains: appChains.chains }),
-  coinbaseWallet({ appName: "scaffold-eth-2", chains: appChains.chains }),
-  rainbowWallet({ chains: appChains.chains }),
+  metaMaskWallet({ ...walletsOptions, shimDisconnect: true }),
+  walletConnectWallet(walletsOptions),
+  ledgerWallet(walletsOptions),
+  braveWallet(walletsOptions),
+  coinbaseWallet({ ...walletsOptions, appName: "scaffold-eth-2" }),
+  rainbowWallet(walletsOptions),
+  ...(!targetNetworks.some(network => network.id !== chains.hardhat.id) || !onlyLocalBurnerWallet
+    ? [
+        burnerWalletConfig({
+          chains: appChains.chains.filter(chain => targetNetworks.map(({ id }) => id).includes(chain.id)),
+        }),
+      ]
+    : []),
+  safeWallet({ ...walletsOptions }),
 ];
 
 /**
@@ -60,6 +71,6 @@ const wallets = [
 export const wagmiConnectors = connectorsForWallets([
   {
     groupName: "Supported Wallets",
-    wallets: burnerConfig.enabled ? [...wallets, burnerWalletConfig({ chains: [appChains.chains[0]] })] : wallets,
+    wallets,
   },
 ]);
