@@ -1,24 +1,17 @@
-import { BlockWithTransactions, TransactionResponse } from "@ethersproject/abstract-provider";
+import { Block, Transaction, TransactionReceipt } from "viem";
 
-export type TransactionWithFunction = TransactionResponse & {
+export type TransactionWithFunction = Transaction & {
   functionName?: string;
   functionArgs?: any[];
   functionArgNames?: string[];
   functionArgTypes?: string[];
 };
 
-export interface Block extends BlockWithTransactions {
-  transactions: TransactionWithFunction[];
-}
+type TransactionReceipts = {
+  [key: string]: TransactionReceipt;
+};
 
-interface TransactionReceipts {
-  [key: string]: {
-    contractAddress: string;
-  };
-}
-
-export interface TransactionsTableProps {
+export type TransactionsTableProps = {
   blocks: Block[];
   transactionReceipts: TransactionReceipts;
-  isLoading: boolean;
-}
+};
