@@ -1,4 +1,3 @@
-import { TransactionReceipt } from "@ethersproject/abstract-provider";
 import {
   Abi,
   AbiParameterToPrimitiveType,
@@ -8,26 +7,70 @@ import {
   ExtractAbiFunction,
 } from "abitype";
 import type { ExtractAbiFunctionNames } from "abitype";
+import type { Simplify } from "type-fest";
+import type { MergeDeepRecord } from "type-fest/source/merge-deep";
+import {
+  Address,
+  Block,
+  GetEventArgs,
+  GetTransactionReceiptReturnType,
+  GetTransactionReturnType,
+  Log,
+  TransactionReceipt,
+} from "viem";
 import { UseContractEventConfig, UseContractReadConfig, UseContractWriteConfig } from "wagmi";
-import contractsData from "~~/generated/deployedContracts";
+import deployedContractsData from "~~/contracts/deployedContracts";
+import externalContractsData from "~~/contracts/externalContracts";
 import scaffoldConfig from "~~/scaffold.config";
 
+type AddExternalFlag<T> = {
+  [ChainId in keyof T]: {
+    [ContractName in keyof T[ChainId]]: T[ChainId][ContractName] & { external?: true };
+  };
+};
+
+const deepMergeContracts = <L extends Record<PropertyKey, any>, E extends Record<PropertyKey, any>>(
+  local: L,
+  external: E,
+) => {
+  const result: Record<PropertyKey, any> = {};
+  const allKeys = Array.from(new Set([...Object.keys(external), ...Object.keys(local)]));
+  for (const key of allKeys) {
+    if (!external[key]) {
+      result[key] = local[key];
+      continue;
+    }
+    const amendedExternal = Object.fromEntries(
+      Object.entries(external[key] as Record<string, Record<string, unknown>>).map(([contractName, declaration]) => [
+        contractName,
+        { ...declaration, external: true },
+      ]),
+    );
+    result[key] = { ...local[key], ...amendedExternal };
+  }
+  return result as MergeDeepRecord<AddExternalFlag<L>, AddExternalFlag<E>, { arrayMergeMode: "replace" }>;
+};
+
+const contractsData = deepMergeContracts(deployedContractsData, externalContractsData);
+
+export type InheritedFunctions = { readonly [key: string]: string };
+
+export type GenericContract = {
+  address: Address;
+  abi: Abi;
+  inheritedFunctions?: InheritedFunctions;
+  external?: true;
+};
+
 export type GenericContractsDeclaration = {
-  [key: number]: readonly {
-    name: string;
-    chainId: string;
-    contracts: {
-      [key: string]: {
-        address: string;
-        abi: Abi;
-      };
-    };
-  }[];
+  [chainId: number]: {
+    [contractName: string]: GenericContract;
+  };
 };
 
 export const contracts = contractsData as GenericContractsDeclaration | null;
 
-type ConfiguredChainId = (typeof scaffoldConfig)["targetNetwork"]["id"];
+type ConfiguredChainId = (typeof scaffoldConfig)["targetNetworks"][0]["id"];
 
 type IsContractDeclarationMissing<TYes, TNo> = typeof contractsData extends { [key in ConfiguredChainId]: any }
   ? TNo
@@ -35,9 +78,7 @@ type IsContractDeclarationMissing<TYes, TNo> = typeof contractsData extends { [k
 
 type ContractsDeclaration = IsContractDeclarationMissing<GenericContractsDeclaration, typeof contractsData>;
 
-export type Chain = keyof ContractsDeclaration;
-
-type Contracts = ContractsDeclaration[ConfiguredChainId][0]["contracts"];
+type Contracts = ContractsDeclaration[ConfiguredChainId];
 
 export type ContractName = keyof Contracts;
 
@@ -63,18 +104,15 @@ export type AbiFunctionOutputs<TAbi extends Abi, TFunctionName extends string> =
 
 export type AbiFunctionReturnType<TAbi extends Abi, TFunctionName extends string> = IsContractDeclarationMissing<
   any,
-  AbiParametersToPrimitiveTypes<AbiFunctionOutputs<TAbi, TFunctionName>>[0]
+  AbiParametersToPrimitiveTypes<AbiFunctionOutputs<TAbi, TFunctionName>> extends readonly [any]
+    ? AbiParametersToPrimitiveTypes<AbiFunctionOutputs<TAbi, TFunctionName>>[0]
+    : AbiParametersToPrimitiveTypes<AbiFunctionOutputs<TAbi, TFunctionName>>
 >;
 
 export type AbiEventInputs<TAbi extends Abi, TEventName extends ExtractAbiEventNames<TAbi>> = ExtractAbiEvent<
   TAbi,
   TEventName
 >["inputs"];
-
-export type AbiEventArgs<
-  TAbi extends Abi,
-  TEventName extends ExtractAbiEventNames<TAbi>,
-> = AbiParametersToPrimitiveTypes<AbiEventInputs<TAbi, TEventName>>;
 
 export enum ContractCodeStatus {
   "LOADING",
@@ -83,25 +121,15 @@ export enum ContractCodeStatus {
 }
 
 type AbiStateMutability = "pure" | "view" | "nonpayable" | "payable";
-
-export type FunctionNamesWithoutInputs<
-  TAbi extends Abi,
-  TAbiStateMutibility extends AbiStateMutability = AbiStateMutability,
-> = Extract<
-  TAbi[number],
-  {
-    type: "function";
-    stateMutability: TAbiStateMutibility;
-    inputs: readonly [];
-  }
->["name"];
+export type ReadAbiStateMutability = "view" | "pure";
+export type WriteAbiStateMutability = "nonpayable" | "payable";
 
 export type FunctionNamesWithInputs<
-  TAbi extends Abi,
+  TContractName extends ContractName,
   TAbiStateMutibility extends AbiStateMutability = AbiStateMutability,
 > = Exclude<
   Extract<
-    TAbi[number],
+    ContractAbi<TContractName>[number],
     {
       type: "function";
       stateMutability: TAbiStateMutibility;
@@ -112,18 +140,6 @@ export type FunctionNamesWithInputs<
   }
 >["name"];
 
-type ReadAbiStateMutability = "view" | "pure";
-type WriteAbiStateMutability = "nonpayable" | "payable";
-
-type RestConfigParam<TAbiStateMutability extends AbiStateMutability> = Partial<
-  Omit<
-    TAbiStateMutability extends ReadAbiStateMutability
-      ? UseContractReadConfig
-      : UseContractWriteConfig<"recklesslyUnprepared">,
-    "chainId" | "abi" | "address" | "functionName" | "args"
-  >
->;
-
 type Expand<T> = T extends object ? (T extends infer O ? { [K in keyof O]: O[K] } : never) : T;
 
 type UnionToIntersection<U> = Expand<(U extends any ? (k: U) => void : never) extends (k: infer I) => void ? I : never>;
@@ -132,9 +148,8 @@ type OptionalTupple<T> = T extends readonly [infer H, ...infer R] ? readonly [H 
 
 type UseScaffoldArgsParam<
   TContractName extends ContractName,
-  TAbiStateMutability extends AbiStateMutability,
-  TFunctionName extends ExtractAbiFunctionNames<ContractAbi<TContractName>, TAbiStateMutability>,
-> = TFunctionName extends FunctionNamesWithInputs<ContractAbi<TContractName>, TAbiStateMutability>
+  TFunctionName extends ExtractAbiFunctionNames<ContractAbi<TContractName>>,
+> = TFunctionName extends FunctionNamesWithInputs<TContractName>
   ? {
       args: OptionalTupple<UnionToIntersection<AbiFunctionArguments<ContractAbi<TContractName>, TFunctionName>>>;
     }
@@ -151,8 +166,8 @@ export type UseScaffoldReadConfig<
   Partial<UseContractReadConfig>,
   {
     functionName: TFunctionName;
-  } & UseScaffoldArgsParam<TContractName, ReadAbiStateMutability, TFunctionName> &
-    RestConfigParam<ReadAbiStateMutability>
+  } & UseScaffoldArgsParam<TContractName, TFunctionName> &
+    Omit<UseContractReadConfig, "chainId" | "abi" | "address" | "functionName" | "args">
 >;
 
 export type UseScaffoldWriteConfig<
@@ -160,29 +175,51 @@ export type UseScaffoldWriteConfig<
   TFunctionName extends ExtractAbiFunctionNames<ContractAbi<TContractName>, WriteAbiStateMutability>,
 > = {
   contractName: TContractName;
-  value?: string;
   onBlockConfirmation?: (txnReceipt: TransactionReceipt) => void;
   blockConfirmations?: number;
 } & IsContractDeclarationMissing<
-  Partial<UseContractWriteConfig<"recklesslyUnprepared">> & { args?: unknown[] },
+  Partial<UseContractWriteConfig>,
   {
     functionName: TFunctionName;
-  } & UseScaffoldArgsParam<TContractName, WriteAbiStateMutability, TFunctionName> &
-    RestConfigParam<WriteAbiStateMutability>
+  } & UseScaffoldArgsParam<TContractName, TFunctionName> &
+    Omit<UseContractWriteConfig, "chainId" | "abi" | "address" | "functionName" | "args" | "mode">
 >;
 
 export type UseScaffoldEventConfig<
   TContractName extends ContractName,
   TEventName extends ExtractAbiEventNames<ContractAbi<TContractName>>,
-  TEventInputs extends AbiEventArgs<ContractAbi<TContractName>, TEventName> & any[],
+  TEvent extends ExtractAbiEvent<ContractAbi<TContractName>, TEventName> = ExtractAbiEvent<
+    ContractAbi<TContractName>,
+    TEventName
+  >,
 > = {
   contractName: TContractName;
 } & IsContractDeclarationMissing<
-  UseContractEventConfig & { listener: (...args: any[]) => void },
-  {
-    eventName: TEventName;
-    listener: (...args: TEventInputs) => void;
-    once?: boolean;
+  Omit<UseContractEventConfig, "listener"> & {
+    listener: (
+      logs: Simplify<
+        Omit<Log<bigint, number, any>, "args" | "eventName"> & {
+          args: Record<string, unknown>;
+          eventName: string;
+        }
+      >[],
+    ) => void;
+  },
+  Omit<UseContractEventConfig<ContractAbi<TContractName>, TEventName>, "listener"> & {
+    listener: (
+      logs: Simplify<
+        Omit<Log<bigint, number, false, TEvent, false, [TEvent], TEventName>, "args"> & {
+          args: AbiParametersToPrimitiveTypes<TEvent["inputs"]> &
+            GetEventArgs<
+              ContractAbi<TContractName>,
+              TEventName,
+              {
+                IndexedOnly: false;
+              }
+            >;
+        }
+      >[],
+    ) => void;
   }
 >;
 
@@ -209,12 +246,47 @@ export type EventFilters<
 export type UseScaffoldEventHistoryConfig<
   TContractName extends ContractName,
   TEventName extends ExtractAbiEventNames<ContractAbi<TContractName>>,
+  TBlockData extends boolean = false,
+  TTransactionData extends boolean = false,
+  TReceiptData extends boolean = false,
 > = {
   contractName: TContractName;
   eventName: IsContractDeclarationMissing<string, TEventName>;
-  fromBlock: number;
+  fromBlock: bigint;
   filters?: EventFilters<TContractName, TEventName>;
-  blockData?: boolean;
-  transactionData?: boolean;
-  receiptData?: boolean;
+  blockData?: TBlockData;
+  transactionData?: TTransactionData;
+  receiptData?: TReceiptData;
+  watch?: boolean;
+  enabled?: boolean;
 };
+
+export type UseScaffoldEventHistoryData<
+  TContractName extends ContractName,
+  TEventName extends ExtractAbiEventNames<ContractAbi<TContractName>>,
+  TBlockData extends boolean = false,
+  TTransactionData extends boolean = false,
+  TReceiptData extends boolean = false,
+  TEvent extends ExtractAbiEvent<ContractAbi<TContractName>, TEventName> = ExtractAbiEvent<
+    ContractAbi<TContractName>,
+    TEventName
+  >,
+> =
+  | IsContractDeclarationMissing<
+      any[],
+      {
+        log: Log<bigint, number, false, TEvent, false, [TEvent], TEventName>;
+        args: AbiParametersToPrimitiveTypes<TEvent["inputs"]> &
+          GetEventArgs<
+            ContractAbi<TContractName>,
+            TEventName,
+            {
+              IndexedOnly: false;
+            }
+          >;
+        block: TBlockData extends true ? Block<bigint, true> : null;
+        receipt: TReceiptData extends true ? GetTransactionReturnType : null;
+        transaction: TTransactionData extends true ? GetTransactionReceiptReturnType : null;
+      }[]
+    >
+  | undefined;
