@@ -1,33 +1,38 @@
+import { TransactionWithFunction } from "./block";
 import { GenericContractsDeclaration } from "./contract";
-import { JsonFragment } from "@ethersproject/abi";
-import { ethers } from "ethers";
-import { hardhat } from "wagmi/chains";
-import contractData from "~~/generated/deployedContracts";
-import { TransactionWithFunction } from "~~/utils/scaffold-eth";
+import { Abi, decodeFunctionData, getAbiItem } from "viem";
+import { hardhat } from "viem/chains";
+import contractData from "~~/contracts/deployedContracts";
 
-type ContractsInterfaces = Record<string, ethers.utils.Interface>;
-type FunctionArgNameType = string;
-type IndexType = number;
+type ContractsInterfaces = Record<string, Abi>;
 type TransactionType = TransactionWithFunction | null;
 
 const deployedContracts = contractData as GenericContractsDeclaration | null;
-const chainMetaData = deployedContracts?.[hardhat.id]?.[0];
+const chainMetaData = deployedContracts?.[hardhat.id];
 const interfaces = chainMetaData
-  ? Object.entries(chainMetaData.contracts).reduce((finalInterfacesObj, [contractName, contract]) => {
-      finalInterfacesObj[contractName] = new ethers.utils.Interface(contract.abi as JsonFragment[]);
+  ? Object.entries(chainMetaData).reduce((finalInterfacesObj, [contractName, contract]) => {
+      finalInterfacesObj[contractName] = contract.abi;
       return finalInterfacesObj;
     }, {} as ContractsInterfaces)
   : {};
 
 export const decodeTransactionData = (tx: TransactionWithFunction) => {
-  if (tx.data.length >= 10) {
-    for (const [, contractInterface] of Object.entries(interfaces)) {
+  if (tx.input.length >= 10 && !tx.input.startsWith("0x60e06040")) {
+    for (const [, contractAbi] of Object.entries(interfaces)) {
       try {
-        const decodedData = contractInterface.parseTransaction({ data: tx.data });
-        tx.functionName = `${decodedData.name}`;
-        tx.functionArgs = Object.values(decodedData.args);
-        tx.functionArgNames = contractInterface.getFunction(decodedData.name).inputs.map((input: any) => input.name);
-        tx.functionArgTypes = contractInterface.getFunction(decodedData.name).inputs.map((input: any) => input.type);
+        const { functionName, args } = decodeFunctionData({
+          abi: contractAbi,
+          data: tx.input,
+        });
+        tx.functionName = functionName;
+        tx.functionArgs = args as any[];
+        tx.functionArgNames = getAbiItem({ abi: contractAbi, name: functionName }).inputs.map(
+          (input: any) => input.name,
+        );
+        tx.functionArgTypes = getAbiItem({ abi: contractAbi, name: functionName }).inputs.map(
+          (input: any) => input.type,
+        );
+
         break;
       } catch (e) {
         console.error(`Parsing failed: ${e}`);
@@ -46,8 +51,7 @@ export const getFunctionDetails = (transaction: TransactionType) => {
     transaction.functionArgs
   ) {
     const details = transaction.functionArgNames.map(
-      (name: FunctionArgNameType, i: IndexType) =>
-        `${transaction.functionArgTypes?.[i] || ""} ${name} = ${transaction.functionArgs?.[i] || ""}`,
+      (name, i) => `${transaction.functionArgTypes?.[i] || ""} ${name} = ${transaction.functionArgs?.[i] ?? ""}`,
     );
     return `${transaction.functionName}(${details.join(", ")})`;
   }
