@@ -1,7 +1,6 @@
 export * from "./fetchPriceFromUniswap";
-export * from "./getLocalProvider";
 export * from "./networks";
 export * from "./notification";
-export * from "./parseTxnValue";
 export * from "./block";
 export * from "./decodeTxData";
+export * from "./getParsedError";
